requirements.txt
aiosqlite==0.15.0
discord.py==1.5.1
git+git://github.com/Rapptz/discord-ext-menus.git
